OverviewPlots.cs: Bug fix -> descr.
Given that we need to plot 2 sessions, with the 1st having 10 sensors and the 2nd having 5. Graph lines were initialized in a List<CurveItem> instead of a List with Lists of CurveItems. Why? Because each session has its own sensors.
In this case, the result would be a List with 14 sensors. And when Plot() tried to AddPoints to every curve, it would first parse the list until the 10th sensor, then set the _sensor_index=0 and re-parse the same list until the 5th sensor. The result is to have a Graph with the the first 5 sensors have both Session 1 & 2 values, the rest 5 sensors have Session 1 values and the actual sensors of the 2nd session only have their initial values