OverviewPlots.cs: Graph hovering finds items in listbox
[To be implemented]
When 2 or more sensors share the same value for a sample, then a secondary listbox is created which contains those samples. This is done. What's left to be done, is to click on the secondary listbox and then make a connection with the main listbox, to show the user which specific sample that was.