MainWindow.cs: [Bug fix]
If a session is started after a Reset, starting the timers will result in a crash due to an "Index out of bounds" Exception, raised inside RefreshBars(). This happens because the function will be called before the Constructor and it will use the old "_sensor_values" from the last session. So, when a Reset happens, the only proper way to start the Plotting Forms' Timers is by calling their constructor.